Luxury bar cruiser (#303)
* Testing for Barrows Luxury Shuttle

A luxury shuttle for the station to enjoy, creating a high roleplay currency based environment.

* Renaming File

Working on renaming the file

* Small organization and shield/removes turrets

removes the guns as they are annoying and replaces them with mining ones. Adds a shield.

* Fixes the newline issues and whitespace

Fixes the whitespace and end on newline errors.

* Last whitespace error

Glad I caught this! One tricky space key