Supercruiser Aggregeddon and Mezzeriah (#341)
* SuperCruiser Aggregeddon

Adds the basics of the Aggregeddon and provides the basis of the ship.

* Updates the map for quality of life

adds things around the ship to make it a bit more lived in. Fleshed out each 'department' zone. Provides a method for teleportation. Adds lights.

* Adds two supercruisers now, Mezzeriah

another double gunboat, this time on silos!

* Deletes bugged File

e

* Adds details and power

Should fix power issues and set up more environment, as well as get the gun bay doors working

* Finishes dirt and doors

All the doors, power, and systems of the ship are functional for the Aggregeddon.

* Makes the ODST FoxCat Suit Not Slowdown

Should encourage less use of the teleportation feature, as it is a softsuit anyway.

* Removes change to different pr

oops!