Slapcraft Framework (#1069)
* Slapcrafting Framework

Adds a Slapcraft Framework, which is a system that easily allows implementation of crafting recipes which rely on slapping items against one another. There is a button next to a crafting menu that opens a handbook with recipes for it.
Currently this comes in with only 2 recipes, but more can be easily added. (IED and Flamethrower are the ones)
There is a HUD element which opens a Handbook with all the recipes inside.
The steps inside the handbook can link to another recipe for easy navigation across crafting components. (Currently none utilize this)

* Handbook, and other progress

* List description

* Radial menu to aid colliding recipes

* Optimizing recipe lookup with cache

* Sanity for the sake of cache

* More progress and IED recipe

* Blacklist and comments

* More control on the recipe level.

* Flamethrower recipe and more system features

* Removes IED and Flamethrower crafting menu recipes

* Reagent list support for reagent and reagent_container step types

* Load steps before recipes

* Wow I need more sleep

* Multiple results out of one recipe, descriptions

* Freeform step order support.

* Updated flamethrower recipe

* Temperature checks support for reagent steps

* Examine hints for recipes on items.

* Maximum volume for reagent container step

* Better IED hint

* Support for optional steps

* Fine here's the icon for other UI styles

* Abstractiveness

Co-authored-by: Alex 'Avunia' Takiya <6952402+atakiya@users.noreply.github.com>